Amend purge_old_redirects management command
Only write to stdout if redirects have actually been purged

+++ redirects/management/commands/purge_old_redirects.py
@@ -8,7 +8,7 @@ class Command(NoArgsCommand):
     def handle_noargs(self,*args,**kwargs):
         now = django.utils.timezone.now()
         queryset = LiveRedirect.objects.filter(expiry__lte=now)
-        count = queryset.count()
-        queryset.delete()
-
-        self.stdout.write('Purged %s out of date LiveRedirects!\n' % count)
+        if queryset.exists():
+            count = queryset.count()
+            queryset.delete()
+            self.stdout.write('Purged %s out of date LiveRedirects!\n' % count)
